+++ test/unit/CollectionArray.cs
@@ -193,5 +193,160 @@ namespace DecTest
             Assert.AreEqual(result.dataEmpty, new[] { 10, 9, 8, 7, 6 });
             Assert.AreEqual(result.dataProvided, new[] { 10, 9, 8, 7, 6 });
         }
+
+        [Test]
+        public void MultiDimensional([ValuesExcept(RecorderMode.Validation)] RecorderMode mode)
+        {
+            var arr = new int[2,2,3] {
+                { { 1, 2, 3 }, { 4, 5, 6 } },
+                { { 7, 8, 9 }, { 10, 11, 12 } }
+            };
+
+            var deserialized = DoRecorderRoundTrip(arr, mode);
+
+            Assert.AreEqual(arr, deserialized);
+        }
+
+        class MultidimDec : Dec.Dec
+        {
+            public int[,,] data = new int[1,1,1] { { { 10 } } };
+        }
+
+        [Test]
+        public void MultiInsufficient([ValuesExcept(ParserMode.Validation)] ParserMode mode)
+        {
+            UpdateTestParameters(new Dec.Config.UnitTestParameters { explicitTypes = new Type[] { typeof(MultidimDec) } });
+
+            var parser = new Dec.Parser();
+            parser.AddString(Dec.Parser.FileType.Xml, @"
+                <Decs>
+                    <MultidimDec decName=""TestDec"">
+                            <data>
+                                <li>
+                                    <li>
+                                        <li>1</li>
+                                        <li>2</li>
+                                        <li>3</li>
+                                    </li>
+                                    <li>
+                                        <li>4</li>
+                                        <li>5</li>
+                                        <li>6</li>
+                                    </li>
+                                </li>
+                                <li>
+                                    <li>
+                                        <li>7</li>
+                                        <li>8</li>
+                                    </li>
+                                    <li>
+                                        <li>10</li>
+                                        <li>11</li>
+                                        <li>12</li>
+                                    </li>
+                                </li>
+                            </data>
+                    </MultidimDec>
+                </Decs>");
+            ExpectErrors(() => parser.Finish());
+
+            DoParserTests(mode);
+
+            var result = Dec.Database<MultidimDec>.Get("TestDec");
+            Assert.IsNotNull(result);
+
+            Assert.AreEqual(new int[2,2,3] {
+                { { 1, 2, 3 }, { 4, 5, 6 } },
+                { { 7, 8, 0 }, { 10, 11, 12 } }
+            }, result.data);
+        }
+
+        [Test]
+        public void MultiExcessive([ValuesExcept(ParserMode.Validation)] ParserMode mode)
+        {
+            UpdateTestParameters(new Dec.Config.UnitTestParameters { explicitTypes = new Type[] { typeof(MultidimDec) } });
+
+            var parser = new Dec.Parser();
+            parser.AddString(Dec.Parser.FileType.Xml, @"
+                <Decs>
+                    <MultidimDec decName=""TestDec"">
+                            <data>
+                                <li>
+                                    <li>
+                                        <li>1</li>
+                                        <li>2</li>
+                                        <li>3</li>
+                                    </li>
+                                    <li>
+                                        <li>4</li>
+                                        <li>5</li>
+                                        <li>6</li>
+                                    </li>
+                                </li>
+                                <li>
+                                    <li>
+                                        <li>7</li>
+                                        <li>8</li>
+                                        <li>9</li>
+                                        <li>42</li>
+                                    </li>
+                                    <li>
+                                        <li>10</li>
+                                        <li>11</li>
+                                        <li>12</li>
+                                    </li>
+                                </li>
+                            </data>
+                    </MultidimDec>
+                </Decs>");
+            ExpectErrors(() => parser.Finish());
+
+            DoParserTests(mode);
+
+            var result = Dec.Database<MultidimDec>.Get("TestDec");
+            Assert.IsNotNull(result);
+
+            Assert.AreEqual(new int[2,2,3] {
+                { { 1, 2, 3 }, { 4, 5, 6 } },
+                { { 7, 8, 9 }, { 10, 11, 12 } }
+            }, result.data);
+        }
+
+        [Test]
+        public void MultiAppend([ValuesExcept(ParserMode.Validation)] ParserMode mode)
+        {
+            UpdateTestParameters(new Dec.Config.UnitTestParameters { explicitTypes = new Type[] { typeof(MultidimDec) } });
+
+            var parser = new Dec.Parser();
+            parser.AddString(Dec.Parser.FileType.Xml, @"
+                <Decs>
+                    <MultidimDec decName=""TestDec"">
+                            <data mode=""append"">
+                                <li>
+                                    <li>
+                                        <li>100</li>
+                                    </li>
+                                </li>
+                                <li>
+                                    <li>
+                                        <li>1000</li>
+                                    </li>
+                                </li>
+                            </data>
+                    </MultidimDec>
+                </Decs>");
+            ExpectErrors(() => parser.Finish());
+
+            DoParserTests(mode);
+
+            var result = Dec.Database<MultidimDec>.Get("TestDec");
+            Assert.IsNotNull(result);
+
+            Assert.AreEqual(new int[3,1,1] {
+                { { 10 } },
+                { { 100 } },
+                { { 1000 } }
+            }, result.data);
+        }
     }
 }